+++ src/Babdelaura/AdminBundle/Form/ArticleType.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Babdelaura\BlogBundle\Form;
+namespace Babdelaura\AdminBundle\Form;
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
