+++ src/Babdelaura/AdminBundle/Resources/views/Article/enregistrerArticle.html.twig
@@ -1,4 +1,4 @@
-{% extends "BabdelauraBlogBundle:Admin:layout.html.twig" %}
+{% extends "BabdelauraBlogBundle::layout.html.twig" %}
 
 {% block content %}
     <h1>Ajout d'un article</h1>
