+++ src/Babdelaura/AdminBundle/Resources/views/Article/listerArticles.html.twig
@@ -1,4 +1,4 @@
-{% extends "BabdelauraBlogBundle:Admin:layout.html.twig" %}
+{% extends "BabdelauraBlogBundle::layout.html.twig" %}
 
 {% block content %}
