+++ src/Babdelaura/AdminBundle/Resources/views/Article/listerArticlesGrille.html.twig
@@ -1,0 +1,5 @@
+{% extends "BabdelauraBlogBundle::layout.html.twig" %}
+{% block content %}
+    <h1>Les derniers articles</h1>
+    {% include "BabdelauraBlogBundle:Article:grille.html.twig" with {'modeAdmin': true} %}
+{% endblock %}
