--- src/Babdelaura/BlogBundle/Resources/views/Admin/Article/listerArticlesGrille.html.twig
@@ -1,5 +1,0 @@
-{% extends "BabdelauraBlogBundle:Admin:layout.html.twig" %}
-{% block content %}
-    <h1>Les derniers articles</h1>    
-    {% include "BabdelauraBlogBundle:Article:grille.html.twig" with {'modeAdmin': true} %}
-{% endblock %}
